requirements.txt
# List your project dependencies here
dotenv
pytest
streamlit
autogen
pyyaml
